backend/requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.12.6
aiohttp-retry==2.9.1
aiosignal==1.3.2
asgiref==3.8.1
attrs==25.3.0
certifi==2025.4.26
chardet==5.2.0
charset-normalizer==3.4.2
dj-database-url==2.3.0
Django==5.2.1
django-cors-headers==4.7.0
djangorestframework==3.16.0
djangorestframework_simplejwt==5.5.0
frozenlist==1.6.0
idna==3.10
multidict==6.4.4
pillow==11.2.1
propcache==0.3.1
psycopg2==2.9.10
psycopg2-binary==2.9.10
PyJWT==2.9.0
python-dotenv==1.1.0
pytz==2025.2
reportlab==4.4.1
requests==2.32.3
sqlparse==0.5.3
twilio==9.6.2
typing_extensions==4.13.2
tzdata==2025.2
urllib3==2.4.0
yarl==1.20.0
